Extract module-related helpers
